rename Blazor_net5.0 to BlazorApp
