+++ src/mono/wasm/testassets/BlazorApp/App.razor
