+++ src/mono/wasm/testassets/BlazorApp/Pages/Counter.razor
