+++ src/mono/wasm/testassets/BlazorApp/Pages/FetchData.razor
