+++ src/mono/wasm/testassets/BlazorApp/Pages/Index.razor
