+++ src/mono/wasm/testassets/BlazorApp/Shared/MainLayout.razor
