+++ src/mono/wasm/testassets/BlazorApp/Shared/NavMenu.razor
@@ -1,5 +1,5 @@
 ﻿<div class="top-row pl-4 navbar navbar-dark">
-    <a class="navbar-brand" href="">Blazor_net50</a>
+    <a class="navbar-brand" href="">BlazorApp</a>
     <button class="navbar-toggler" @onclick="ToggleNavMenu">
         <span class="navbar-toggler-icon"></span>
     </button>
