+++ src/mono/wasm/testassets/BlazorApp/Shared/SurveyPrompt.razor
