+++ src/mono/wasm/testassets/BlazorApp/_Imports.razor
@@ -6,5 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using Blazor_net50
-@using Blazor_net50.Shared
+@using BlazorApp
+@using BlazorApp.Shared
